requirements.txt
flask==3.1.1
flask-cors==6.0.1
flask-socketio==5.5.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
requests==2.32.4
werkzeug==3.1.3

